Revert "Use CMake for Full Build and add CI (#17)"
This reverts commit f2531476f4b07c241358a282fd23b0b95beaa433.
