+++ src/main.hpp
@@ -1,7 +1,6 @@
-#include <iostream>
-
 #include "mesh.hpp"
 #include "plasma.hpp"
 
+int main();
 void initialize();
 void evolve(Mesh *mesh, Plasma *plasma);
