+++ test/CMakeLists.txt
@@ -1,26 +1,35 @@
-include(FetchContent)
+cmake_minimum_required(VERSION 3.14)
+project(PolyrepoPracticeCore VERSION 0.0.1)
 
+# GoogleTest requires at least C++11
+set(CMAKE_CXX_STANDARD 11)
+
+include(FetchContent)
 FetchContent_Declare(
   googletest
   URL https://github.com/google/googletest/archive/054a986a8513149e8374fc669a5fe40117ca6b41.zip
 )
-
 # For Windows: Prevent overriding the parent project's compiler/linker settings
 set(gtest_force_shared_crt ON CACHE BOOL "" FORCE)
 FetchContent_MakeAvailable(googletest)
 
 enable_testing()
 
-file( GLOB TEST_SRCS ${CMAKE_CURRENT_SOURCE_DIR}/test_*.cpp )
+add_executable(
+  mesh_test
+  mesh_test.cc
+  plasma_test
+  plasma_test.cc
+  ../src/mesh.cpp
+  ../src/plasma.cpp
+)
+target_link_libraries(
+  mesh_test
+  #plasma_test
+  gtest_main
+  lapack
+)
 
 include(GoogleTest)
-
-foreach( TEST ${TEST_SRCS} )
-  get_filename_component( EXECUTABLE ${TEST} NAME_WLE )
-  message( STATUS "Found test - ${EXECUTABLE}" )
-  add_executable( ${EXECUTABLE} ${TEST} )
-  target_link_libraries( ${EXECUTABLE} ${LIBRARY_NAME} gtest_main lapack )
-  target_include_directories( ${EXECUTABLE} PRIVATE ${CMAKE_SOURCE_DIR}/include )
-  gtest_discover_tests( ${EXECUTABLE} )
-endforeach()
+gtest_discover_tests(mesh_test plasma_test)
